requirements.txt
streamlit
python-dotenv
pandas
pydantic
supabase==2.0.0
httpx==0.24.1

# LangChain and AI dependencies - Compatible versions
langchain==0.3.7
langchain-core==0.3.15
langchain-community==0.3.5
langchain-google-genai==2.0.5
google-generativeai==0.8.3
langsmith==0.1.0

# Document processing dependencies
PyPDF2
python-docx